Add unit tests for DrsResolutionService.fetchObjectInfo
Note: using @VisibleForTesting can be a code smell, but in this case there were no existing tests of DrsResolutionService and the code's public methods were not easy to unit test.
I decided that it was better to add targeted unit test coverage for this time-sensitive bugfix than rush a refactor.