[DR-3330] DrsResolutionServiceTest needn't be a SpringBootTest
By downgrading to a non-Spring Boot test that uses mocks with strict stubbing, we simplify the test set-up and make it less likely that future expansion of DrsResolutionService results in live dependencies being inadvertently loaded in the test context.
The strict stubbing also allows us to remove verifications of actions that we've mocked: if our test didn't perform the action, strict stubbing would fail the test.