Expand TrackingInterceptorTest to ensure that disabling Bard emissions works
The test previously used a live DrsHubConfig: I refactored slightly to mock it so that we could manipulate it.
Noting that this unit test is more of a functional test that tests many components at once.
With more time I might be inclined to break up the testing into different unit test suites, but this does offer us fast, not-flaky coverage so I'll leave it as is for now.