Spying not needed on RestTemplateFactory in DrsApiFactoryTest
But, we no longer have actual invocation of these RestTemplateFactory methods in our tests.  Writing isolated unit tests for them don't make sense.
Unsure of the best approach here, it may instead be to treat DrsApiFactoryTest as a sociable unit test whcich spies RestTemplateFactory to verify caching behavior but does not mock it.