DR-3385 CVE-2023-6378 Upgrade logback-classic, logback-core to 1.4.14
It was not enough that we updated TCL, since we exclude its spring dependency.
Spring boot hasn't yet upgraded its logback versions, so for the time being we can pull in non-vulnerable versions ourselves.
Added a note to check back on this following the next Spring boot release.

While here, realized that I wasn't pulling in the latest 3.1.x Spring boot version, so brought it up to 3.1.6.